backend/requirements-dev.txt
-r requirements.txt

# ── Dev / debug ───────────────────────────────────────────────────────────────
django-debug-toolbar==4.4.2
django-extensions==3.2.3
ipython==8.26.0
Werkzeug==3.0.3

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.1

# ── Code quality ──────────────────────────────────────────────────────────────
black==24.8.0
isort==5.13.2
flake8==7.1.1
